+++ apps/core-lib-dev/src/app/app.element.ts
@@ -5,10 +5,10 @@ import {
   GdsOption,
   htmlTemplateTagFactory,
   getScopedTagName,
-  GdsFormExp,
 } from '@sebgroup/green-core'
 
-import { registerTransitionalStyles, setLocale } from '@sebgroup/green-core'
+import { registerTransitionalStyles } from '@sebgroup/green-core/transitional-styles'
+import { setLocale } from '@sebgroup/green-core/localization'
 
 const html = htmlTemplateTagFactory((strs: TemplateStringsArray) =>
   strs.join('')
@@ -29,6 +29,7 @@ export class AppElement extends HTMLElement {
         <div class="card">
           <form id="my-form">
             <gds-form-validator></gds-form-validator>
+
             <gds-dropdown
               searchable
               multiple
@@ -49,6 +50,7 @@ export class AppElement extends HTMLElement {
             <input name="my-text" />
 
             <button type="submit">Submit</button>
+            <button type="reset">Reset</button>
           </form>
         </div>
       </div>
@@ -57,7 +59,6 @@ export class AppElement extends HTMLElement {
     const dropdown = this.querySelector('#dropdown') as GdsDropdown
 
     dropdown.addEventListener('change', (e: CustomEvent) => {
-      console.log('Is Droopdown? ', e.target instanceof GdsDropdown)
       console.log('Value changed: ', e.detail.value)
       console.log('dropdown', dropdown.options)
       console.log('dd name', dropdown.name)