+++ jest.preset.js
@@ -12,4 +12,6 @@ module.exports = {
    * More info: https://jestjs.io/docs/upgrading-to-jest29#snapshot-format
    */
   snapshotFormat: { escapeString: true, printBasicPrototype: true },
+
+  setupFilesAfterEnv: [__dirname + '/setup-jest.js'],
 }