+++ libs/core/src/utils/testing/index.ts
@@ -7,8 +7,8 @@ function determineMousePosition(
   offsetY: number
 ) {
   const { x, y, width, height } = el.getBoundingClientRect()
-  const centerX = Math.floor(x + window.pageXOffset + width / 2)
-  const centerY = Math.floor(y + window.pageYOffset + height / 2)
+  const centerX = Math.floor(x + window.scrollX + width / 2)
+  const centerY = Math.floor(y + window.scrollY + height / 2)
   let clickX: number
   let clickY: number