+++ libs/react/src/lib/form/input/input.tsx
@@ -16,9 +16,9 @@ export type Renderer = (
   onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void,
   onChangeInput?: (value: string) => string,
   label?: string,
-  info?: string,
+  info?: string | React.ReactNode,
   validator?: IValidator,
-  expandableInfo?: string,
+  expandableInfo?: React.ReactNode,
   expandableInfoButtonLabel?: string,
   testId?: string
 ) => JSX.Element