+++ libs/react/src/lib/tabs/tabs.spec.tsx
@@ -1,7 +1,6 @@
 import { render, screen, fireEvent } from '@testing-library/react'
 import Tabs, { IList, Tab } from './tabs'
 
-
 const list: IList[] = [
   { text: 'Page 1', href: '#' },
   { text: 'Page 2', href: '#' },
@@ -72,6 +71,40 @@ describe('Tabs allow components as content', () => {
       expect(tag.textContent).toBe(tabList[index].title);
     }  
     )
+  })
+
+  it('should render the tab set as selected by default', () => {
+    render(<Tabs>
+        <Tab title='NotSelected'></Tab>
+        <Tab title='NotSelected'></Tab>
+        <Tab title='Selected' selected>Tab is selected</Tab>
+      </Tabs>)
+    const tabsContent: HTMLElement[] = screen.getAllByRole('tabpanel')
+
+    expect(tabsContent).toBeTruthy()
+    expect(tabsContent[0].textContent).toBe("Tab is selected"); 
+  })
+  
+  it('should only render the selected tab element', () => {
+    render(<Tabs>{ tabList.map(tab => <Tab {...tab}>{tab.children}</Tab>)}</Tabs>)
+    const tabsContent: HTMLElement[] = screen.getAllByRole('tabpanel', { hidden: true})
+    const anchorTag: HTMLAnchorElement[] = screen.getAllByRole('tab')
+    
+    expect(anchorTag).toBeTruthy()
+    expect(tabsContent.length).toEqual(6) 
+    expect(tabsContent[0].textContent).toBe(tabList[0].children);
+    tabsContent?.forEach((tab, index) => {
+      if(index !== 0)
+        expect(tab.textContent).toBe("");
+    })
+    const clickedTab = 2
+    fireEvent.click(anchorTag[clickedTab])
+    expect(tabsContent[clickedTab].textContent).toBe(tabList[clickedTab].children);
+
+    tabsContent?.forEach((tab, index) => {
+      if(index !== clickedTab)
+        expect(tab.textContent).toBe("");
+    })
   })
 
   it('onClick changes selectedTab', () => {